+++ st2api/st2api/controllers/v1/packs.py
@@ -103,6 +103,9 @@ class PackInstallController(ActionExecutionsControllerMixin):
         if pack_install_request.force:
             parameters['force'] = True
 
+        if pack_install_request.skip:
+            parameters['skip'] = True
+
         if not requester_user:
             requester_user = UserDB(cfg.CONF.system_user.user)