+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ Added
 * Add support for blacklisting / whitelisting hosts to the HTTP runner by adding new
   ``url_hosts_blacklist`` and ``url_hosts_whitelist`` runner attribute. (new feature)
   #4757
+* Add support for Pack dependency management to install dependency packs and find out confilct packs.
+  #4769
 
 Changed
 ~~~~~~~